requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Validation and Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# File uploads
python-multipart>=0.0.6

# Database (Story 1.2)
sqlalchemy>=2.0.0
# alembic>=1.12.0  # Uncomment if migrations needed

# AI Integration (Story 2.1)
google-genai>=1.0.0
pillow>=10.0.0

# OpenAI (Principal Designer AI - GPT-4o Vision)
openai>=1.0.0

# Security (Story 1.3)
# python-magic>=0.4.27  # File type detection

# HTTP Client
httpx>=0.25.0

# Image processing
pillow>=10.1.0
